Extracting interface for logout service
